RD-2690 Fix scrolling form to top when there are errors
When there are new errors in the form, the functionality for scrolling
the form back to the top to see them was broken since at least 2019-11.

The problem was that the DOM node for `this` cound not have been found
in a function component.

The fix is to use a ref to get a handle to the form and then invoke
scrolling on that.

There is also a new test that makes sure the scroll method is invoked.